added remove cell
